requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
bson==0.5.10
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
decorator==5.1.0
feedparser==6.0.8
joblib==1.0.1
jsonpath-ng==1.5.3
nltk==3.6.2
paramiko==2.7.2
ply==3.11
pycparser==2.20
pymongo==3.12.0
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
regex==2021.8.3
selenium==3.141.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
sshtunnel==0.4.0
tqdm==4.62.0
urllib3==1.26.6
